Bugs fixed + prettier
I know, projectgroup needs to be Eloquent Friendly, but thats an other issue

+++ app/Http/Controllers/ProjectGroupController.php
@@ -6,41 +6,55 @@ use App\Http\Requests\ProjectgroupRequest;
 use App\Models\Address;
 use App\Models\Company;
 use App\Models\contact\Contact;
+use App\Models\contact\ContactType;
+use App\Models\contact\Gender;
 use App\Models\Project;
-use App\Models\ProjectGroup;
+use App\Models\Projectgroup;
+use App\Models\StudentClass;
 use App\Models\User;
-use Exception;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
-use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
-class ProjectGroupController extends Controller
+class ProjectgroupController extends Controller
 {
   /**
    * Display a listing of the resource.
    *
-   * @return Application|Factory|View
+   * @return Response
    */
   public function index()
   {
     $projectgroups = [];
 
-    foreach (ProjectGroup::all() as $projectgroup) {
-      $teachers = $projectgroup
-        ->users()
-        ->role('Teacher')
+    foreach (Projectgroup::all() as $projectgroup) {
+      $assigned_to_group = DB::table('project_group_user')
+        ->select('user_id')
+        ->where('project_group_id', $projectgroup->id)
+        ->pluck('user_id');
+
+      $teachers = User::whereIn('id', $assigned_to_group)
+        ->whereHas('roles', function ($q) {
+          $q->where('name', 'teacher');
+        })
         ->get();
 
-      $students = $projectgroup
-        ->users()
-        ->role('Student')
+      $students = User::whereIn('id', $assigned_to_group)
+        ->whereHas('roles', function ($q) {
+          $q->where('name', 'student');
+        })
         ->get();
 
       $project = Project::find($projectgroup->project);
 
-      $projectname = $project != null ? $project->name : 'Geen Project';
+      $projectname = 'Geen Project';
+      if ($project != null) {
+        $projectname = $project->name;
+      }
 
       array_push($projectgroups, [
         'group' => $projectgroup,
@@ -56,16 +70,25 @@ class ProjectGroupController extends Controller
   private function addClassToStudent($students)
   {
     foreach ($students as $student) {
-      $className = $student->classrooms()->first()->name ?? 'Geen Klas';
+      $student_has_class_room = DB::table('student_class_user')
+        ->where('user_id', '=', $student->id)
+        ->first();
 
-      $student->classroom = $className;
+      if (is_null($student_has_class_room)) {
+        $student->classroom = 'Geen Klas';
+        continue;
+      }
+
+      $class = StudentClass::where('id', $student_has_class_room->student_class_id)->first();
+
+      $student->classroom = $class->name;
     }
   }
 
   /**
    * Show the form for creating a new resource.
    *
-   * @return Application|Factory|View
+   * @return Response
    */
   public function create()
   {
@@ -95,30 +118,41 @@ class ProjectGroupController extends Controller
   /**
    * Store a newly created resource in storage.
    *
-   * @param ProjectgroupRequest $request
-   * @return RedirectResponse
+   * @param Request $request
+   * @return Response
    */
-  public function store(ProjectgroupRequest $request): RedirectResponse
+  public function store(ProjectgroupRequest $request)
   {
     $request->validated();
 
-    $projectGroup = new Projectgroup();
-    $projectGroup->name = $request->name;
+    $group = new Projectgroup();
+    $group->name = $request->name;
 
     if ($request->project != -1) {
-      $projectGroup->project = $request->project;
+      $group->project = $request->project;
     }
 
-    $projectGroup->save();
+    $group->save();
+    $id = $group->id;
 
     // fill in all the users (students and teachers)
     if (isset($request->assignedUsers)) {
-      $projectGroup->users()->sync($request->assignedUsers);
+      foreach ($request->assignedUsers as $assignedUser) {
+        DB::table('project_group_user')->insert([
+          'user_id' => $assignedUser,
+          'project_group_id' => $id,
+        ]);
+      }
     }
 
     // fill in all the contactpersons
     if (isset($request->assignedContacts)) {
-      $projectGroup->contacts()->sync($request->assignedContacts);
+      foreach ($request->assignedContacts as $assignedContact) {
+        DB::table('contact_project_group')->insert([
+          'contact_id' => $assignedContact,
+          'project_group_id' => $id,
+        ]);
+      }
     }
 
     return redirect()->route('projectgroup.index');
@@ -126,28 +160,36 @@ class ProjectGroupController extends Controller
 
   public function show(Projectgroup $projectgroup)
   {
-    $assignedUsers = $projectgroup
-      ->users()
-      ->get()
+    $assignedUsers = DB::table('project_group_user')
+      ->where('project_group_id', '=', $projectgroup->id)
+      ->join('users', 'project_group_user.user_id', '=', 'users.id')
+      ->get('users.id')
       ->pluck('id');
 
-    $students = User::role('Student')
+    $students = User::role('student')
       ->whereIn('id', $assignedUsers)
       ->get();
 
-    $teachers = User::role('Teacher')
+    $teachers = User::role('teacher')
       ->whereIn('id', $assignedUsers)
       ->get();
 
-    $project = Project::all()
-      ->where('id', '=', $projectgroup->project)
+    $project = DB::table('projects')
+      ->where('id', '=', $projectgroup->id)
+      ->get()
       ->first();
 
-    $contacts = $projectgroup->contacts()->get();
+    $assignedContacts = DB::table('contact_project_group')
+      ->where('project_group_id', '=', $projectgroup->id)
+      ->join('contacts', 'contact_project_group.contact_id', '=', 'contacts.id')
+      ->get('contacts.id')
+      ->pluck('id');
+
+    $contacts = Contact::all()->whereIn('id', $assignedContacts);
 
     $contacts = self::matchAdressWithContacts($contacts);
 
-    $newContacts = Contact::all()->wherenotin('id', $contacts->pluck('id'));
+    $newContacts = Contact::all()->wherenotin('id', $assignedContacts);
 
     foreach ($newContacts as $newContact) {
       $newContact->company = [];
@@ -172,23 +214,20 @@ class ProjectGroupController extends Controller
 
   public function addcontact($projectgroupid, $contactid)
   {
-    $projectGroup = Projectgroup::all()
-      ->where('id', '=', $projectgroupid)
-      ->first();
-    try {
-      $projectGroup->contacts()->attach($contactid);
-    } catch (Exception $ex) {
-    }
+    DB::table('projectgroup_has_contacts')->insert([
+      'projectgroupid' => $projectgroupid,
+      'contactid' => $contactid,
+    ]);
 
     return redirect()->route('projectgroup.show', [$projectgroupid]);
   }
 
   public function removecontact($projectgroupid, $contactid)
   {
-    $projectGroup = Projectgroup::all()
-      ->where('id', '=', $projectgroupid)
-      ->first();
-    $projectGroup->contacts()->detach($contactid);
+    DB::table('contact_project_group')
+      ->where('project_group_id', '=', $projectgroupid)
+      ->where('contact_id', '=', $contactid)
+      ->delete();
 
     return redirect()->route('projectgroup.show', [$projectgroupid]);
   }
@@ -202,30 +241,40 @@ class ProjectGroupController extends Controller
   public function edit(Projectgroup $projectgroup)
   {
     $students = User::role('student')->get();
+    $teachers = User::role('teacher')->get();
+    $contacts = Contact::all();
+    $projects = Project::all();
+
+    $assignedUsers =
+      DB::table('project_group_user')
+        ->where('project_group_id', '=', $projectgroup->id)
+        ->join('users', 'project_group_user.user_id', '=', 'users.id')
+        ->get('users.id') ?? [];
+
+    $assignedUsers = array_map(function ($teacher) {
+      return $teacher->id;
+    }, json_decode($assignedUsers));
+
+    $assignedContacts =
+      DB::table('contact_project_group')
+        ->where('project_group_id', '=', $projectgroup->id)
+        ->join('contacts', 'contact_project_group.contact_id', '=', 'contacts.id')
+        ->get('contacts.id') ?? [];
+
+    $assignedContacts = array_map(function ($contact) {
+      return $contact->id;
+    }, json_decode($assignedContacts));
+
     $this->addClassToStudent($students);
 
     return view('projectgroup.manage')
       ->with('projectgroup', $projectgroup)
-      ->with('teachers', User::role('teacher')->get())
-      ->with('students', User::role('student')->get())
-      ->with('contacts', Contact::all())
-      ->with('projects', Project::all())
-      ->with(
-        'assignedUsers',
-        $projectgroup
-          ->users()
-          ->get()
-          ->pluck('id')
-          ->toArray()
-      )
-      ->with(
-        'assignedContacts',
-        $projectgroup
-          ->contacts()
-          ->get()
-          ->pluck('id')
-          ->toArray()
-      )
+      ->with('teachers', $teachers)
+      ->with('students', $students)
+      ->with('contacts', $contacts)
+      ->with('assignedUsers', $assignedUsers)
+      ->with('assignedContacts', $assignedContacts)
+      ->with('projects', $projects)
       ->with('action', 'update');
   }
 
@@ -242,12 +291,68 @@ class ProjectGroupController extends Controller
 
     // insert the connections with users (students and teachers)
     if (isset($request->assignedUsers)) {
-      $projectgroup->users()->sync($request->assignedUsers);
+      foreach ($request->assignedUsers as $assignedUser) {
+        if (
+          !DB::table('project_group_user')
+            ->where('user_id', $assignedUser)
+            ->where('project_group_id', $projectgroup->id)
+            ->exists()
+        ) {
+          DB::table('project_group_user')->insert([
+            'user_id' => $assignedUser,
+            'project_group_id' => $projectgroup->id,
+          ]);
+        }
+      }
     }
 
     // insert the connections with contactpersons
     if (isset($request->assignedContacts)) {
-      $projectgroup->contacts()->sync($request->assignedContacts);
+      foreach ($request->assignedContacts as $assignedContact) {
+        if (
+          !DB::table('project_group_user')
+            ->where('contact_id', $assignedContact)
+            ->where('project_group_id', $projectgroup->id)
+            ->exists()
+        ) {
+          DB::table('contact_project_group')->insert([
+            'contact_id' => $assignedContact,
+            'project_group_id' => $projectgroup->id,
+          ]);
+        }
+      }
+    }
+
+    // removing unnecessary connections with users (students and teachers
+    foreach (
+      DB::table('project_group_user')
+        ->where('project_group_id', $projectgroup->id)
+        ->get()
+      as $dbvalue
+    ) {
+      if (!isset($request->assignedUsers) || !in_array($dbvalue->userid, $request->assignedUsers)) {
+        //DB::delete('DELETE FROM projectgroup_has_users WHERE userid = ? AND projectgroupid = ?', [$dbvalue->userid, $dbvalue->projectgroupid]);
+        DB::table('project_group_user')
+          ->where('user_id', $dbvalue->userid)
+          ->where('project_group_id', $dbvalue->projectgroupid)
+          ->delete();
+      }
+    }
+
+    // removing unnecessary connections with contactpersons
+    foreach (
+      DB::table('contact_project_group')
+        ->where('project_group_id', $projectgroup->id)
+        ->get()
+      as $dbvalue
+    ) {
+      if (!isset($request->assignedContacts) || !in_array($dbvalue->contactid, $request->assignedContacts)) {
+        //DB::delete('DELETE FROM projectgroup_has_contacts WHERE contactid = ? AND projectgroupid = ?', [$dbvalue->contactid, $dbvalue->projectgroupid]);
+        DB::table('contact_project_group')
+          ->where('contact_id', $dbvalue->contactid)
+          ->where('project_group_id', $dbvalue->projectgroupid)
+          ->delete();
+      }
     }
 
     $projectgroup->name = $request->name;
@@ -269,7 +374,9 @@ class ProjectGroupController extends Controller
    */
   public function destroy(Projectgroup $projectgroup)
   {
-    $projectgroup->delete();
+    if (Auth::user()->isAdmin()) {
+      $projectgroup->delete();
+    }
     return redirect()->route('projectgroup.index');
   }
